AbstractRouter<Output, Params>
closes #1
