+++ gradle.properties
@@ -16,7 +16,7 @@
 # This option should only be used with decoupled projects. More details, visit
 # http://www.gradle.org/docs/current/userguide/multi_project_builds.html#sec:decoupled_projects
 # org.gradle.parallel=true
-VERSION_NAME=0.0.5-SNAPSHOT
+VERSION_NAME=0.0.6-SNAPSHOT
 VERSION_CODE=1
 GROUP=com.eyeem.router
